+++ projects/frontend/es-core/src/components/AnimatedImageComponent.cpp
@@ -59,7 +59,7 @@ void AnimatedImageComponent::update(int deltaTime)
 	{
 		mCurrentFrame++;
 
-		if(mCurrentFrame == mFrames.size())
+		if(mCurrentFrame == (int)mFrames.size())
 		{
 			if(mLoop)
 			{
